Max health and armor workaround
player_manager.lua in gmod sets the player's max health on spawn. I could not find a way to modify it, so there is a workaround where we keep track of the player's max health and set it with a slight delay after spawning.

Also sets health and armor when you change the maximums, for convenience.